+++ src/components/ConnectionTable.tsx
@@ -8,10 +8,11 @@ import { useSortBy, useTable } from 'react-table';
 
 import prettyBytes from '../misc/pretty-bytes';
 import s from './ConnectionTable.module.scss';
+import { MutableConnRefCtx } from './conns/ConnCtx';
 
 const sortDescFirst = true;
 
-const columns = [
+const fullColumns = [
   { accessor: 'id', show: false },
   { Header: 'c_host', accessor: 'host' },
   { Header: 'c_process', accessor: 'process' },
@@ -26,6 +27,9 @@ const columns = [
   { Header: 'c_destination_ip', accessor: 'destinationIP' },
   { Header: 'c_type', accessor: 'type' },
 ];
+
+const columns = fullColumns;
+const columnsWithoutProcess = fullColumns.filter((item) => item.accessor !== 'process');
 
 function renderCell(cell: { column: { id: string }; value: number }, locale: Locale) {
   switch (cell.column.id) {
@@ -52,9 +56,10 @@ const tableState = {
 };
 
 function Table({ data }) {
+  const connCtx = React.useContext(MutableConnRefCtx);
   const { getTableProps, headerGroups, rows, prepareRow } = useTable(
     {
-      columns,
+      columns: connCtx.hasProcessPath ? columns : columnsWithoutProcess,
       data,
       initialState: tableState,
       autoResetSortBy: false,
@@ -66,7 +71,13 @@ function Table({ data }) {
   const locale = i18n.language === 'zh' ? zhCN : enUS;
 
   return (
-    <div {...getTableProps()}>
+    <div
+      {...getTableProps()}
+      style={{
+        // @ts-ignore
+        '--col-count': connCtx.hasProcessPath ? '12' : '11',
+      }}
+    >
       {headerGroups.map((headerGroup) => {
         return (
           <div {...headerGroup.getHeaderGroupProps()} className={s.tr}>
@@ -92,7 +103,13 @@ function Table({ data }) {
                     className={cx(
                       s.td,
                       i % 2 === 0 ? s.odd : false,
-                      j >= 2 && j <= 5 ? s.du : false
+                      connCtx.hasProcessPath
+                        ? j >= 2 && j <= 5
+                          ? s.du
+                          : false
+                        : j >= 1 && j <= 4
+                        ? s.du
+                        : false
                     )}
                   >
                     {renderCell(cell, locale)}